Install ms edge versions using single plugin
Ask the user via the plugin (fixes #78)
